ENH: Make sure that we find the proper symbol and not the one that start with _. STYLE: Remove an old style cast

+++ Source/kwsys/DynamicLoader.cxx
@@ -305,7 +305,7 @@ int DynamicLoader::CloseLibrary(LibHandle lib)
   if (lib)
     {
     // The function dlclose() returns 0 on success, and non-zero on error.
-    return !(int)dlclose(lib);
+    return !dlclose(lib);
     }
   // else
   return 0;
